move qbo types into separate file
