+++ src/pages/api/auth/[...nextauth].ts
@@ -4,7 +4,8 @@ import { OAuthConfig } from "next-auth/providers"
 import { CallbacksOptions } from "next-auth"
 
 import { user as firestoreUser } from "@/lib/db"
-import { QBOProfile, getCompanyInfo, OpenIdUserInfo } from "@/lib/qbo-api"
+import { getCompanyInfo } from "@/lib/qbo-api"
+import { QBOProfile, OpenIdUserInfo } from "@/types/qbo-api"
 import { fetchJsonData, base64EncodeString } from "@/lib/util/request"
 import { config } from "@/lib/util/config"
 import { ApiError } from "next/dist/server/api-utils"
