+++ src/pages/api/email.ts
@@ -14,7 +14,8 @@ import { receiptReady } from "@/lib/db-helper"
 import { assertSessionIsQboConnected } from "@/lib/util/next-auth-helper"
 import { WithBody, ReceiptPdfDocument } from "@/components/receipt"
 import { user } from "@/lib/db"
-import { Donation, getDonations } from "@/lib/qbo-api"
+import { getDonations } from "@/lib/qbo-api"
+import { Donation } from "@/types/qbo-api"
 import { getThisYear } from "@/lib/util/date"
 import { downloadImagesForDonee } from "@/lib/db-helper"
 import { config } from "@/lib/util/config"
