+++ src/pages/generate-receipts.tsx
@@ -9,7 +9,8 @@ import { authOptions } from "./api/auth/[...nextauth]"
 import { ReceiptPdfDocument } from "@/components/receipt"
 import { Svg, Link, buttonStyling } from "@/components/ui"
 import { PDFDownloadLink, PDFViewer } from "@/lib/pdfviewer"
-import { Donation, getDonations } from "@/lib/qbo-api"
+import { getDonations } from "@/lib/qbo-api"
+import { Donation } from "@/types/qbo-api"
 import { user } from "@/lib/db"
 import { alreadyFilledIn, receiptReady } from "@/lib/db-helper"
 import { subscribe } from "@/lib/util/request"
